+++ src/components/organisms/Header/Header.tsx
@@ -1,45 +1,24 @@
-import { useLocation, useNavigate } from "react-router-dom";
+import { useAuth } from "../../../hooks/useAuth";
+import { useModal } from "../../../hooks/useModal";
+import { useNav } from "../../../hooks/useNav";
 import { Logo } from "../../atoms/Logo/Logo";
 import { HeaderList } from "../../molecules/HeaderList/HeaderList";
 
 import S from "./Header.module.scss";
-import {
-  useAppDispatch,
-  useAppSelector,
-} from "../../../app/reducers/reduxHooks";
-import { openModal } from "../../../app/reducers/modalSlice";
-import { logout } from "../../../app/reducers/authSlice";
+import { useHeader } from "./useHeader";
 
 export const Header = () => {
-  const dispatch = useAppDispatch();
-  const { isAuth } = useAppSelector((state) => state.auth);
-  const openCreateModal = () => {
-    dispatch(openModal("create"));
-  };
-
-  const handleLogout = () => {
-    dispatch(logout());
-  };
-  const { pathname } = useLocation();
-  const isMainPage = pathname === "/";
-
-  const elements = [
-    { to: "/", text: "Главная" },
-    { to: "/tickets", text: "История" },
-  ];
-
-  const navigate = useNavigate();
-
-  const goToLogin = () => {
-    navigate("/login");
-  };
+  const { links, isAuth, isMainPage } = useHeader();
+  const { goToLogin } = useNav();
+  const { openCreateModal } = useModal();
+  const { toLogOut } = useAuth();
 
   return (
     <header className={S.header}>
       <div className={S.logo}>
         <Logo />
       </div>
-      {isAuth && <HeaderList elements={elements} />}
+      {isAuth && <HeaderList elements={links} />}
 
       {isMainPage && (
         <div onClick={openCreateModal} className={S.logout}>
@@ -47,7 +26,7 @@ export const Header = () => {
         </div>
       )}
       {isAuth ? (
-        <div onClick={handleLogout} className={S.logout}>
+        <div onClick={toLogOut} className={S.logout}>
           выход
         </div>
       ) : (
